pyproject.toml
[project]
name = "giddyanne"
version = "1.5.0"
description = "File watcher with semantic search via embeddings"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0.0",
    "sentence-transformers>=3.0.0",
    "lancedb>=0.15.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyarrow>=18.0.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0.0",
    "pathspec>=0.12.0",
    "tree-sitter-language-pack>=0.13.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
